fix tests for build
